Extract helper to create selections.
