+++ model/selectionHelpers.js
@@ -1,4 +1,4 @@
-import last from '../util/last'
+import { _isDefined, isNil, last } from '../util'
 import Selection from './Selection'
 import PropertySelection from './PropertySelection'
 import ContainerSelection from './ContainerSelection'
@@ -6,6 +6,8 @@ import NodeSelection from './NodeSelection'
 import CustomSelection from './CustomSelection'
 import getContainerRoot from './_getContainerRoot'
 import getContainerPosition from './_getContainerPosition'
+import compareCoordinates from './_compareCoordinates'
+import Coordinate from './Coordinate'
 
 export function fromJSON (json) {
   if (!json) return Selection.nullSelection
@@ -122,6 +124,93 @@ export function setCursor (tx, node, containerPath, mode) {
 
 export function selectNode (tx, nodeId, containerPath) {
   tx.setSelection(createNodeSelection({ doc: tx, nodeId, containerPath }))
+}
+
+export function createSelection (doc, data) {
+  let sel
+  if (isNil(data)) return Selection.nullSelection
+  switch (data.type) {
+    case 'property': {
+      if (isNil(data.endOffset)) {
+        data.endOffset = data.startOffset
+      }
+      if (!_isDefined(data.reverse)) {
+        if (data.startOffset > data.endOffset) {
+          [data.startOffset, data.endOffset] = [data.endOffset, data.startOffset]
+          data.reverse = !data.reverse
+        }
+      }
+      // integrity checks:
+      const text = doc.get(data.path, 'strict')
+      if (data.startOffset < 0 || data.startOffset > text.length) {
+        throw new Error('Invalid startOffset: target property has length ' + text.length + ', given startOffset is ' + data.startOffset)
+      }
+      if (data.endOffset < 0 || data.endOffset > text.length) {
+        throw new Error('Invalid startOffset: target property has length ' + text.length + ', given endOffset is ' + data.endOffset)
+      }
+      sel = new PropertySelection(data)
+      break
+    }
+    case 'container': {
+      const containerPath = data.containerPath
+      const ids = doc.get(containerPath)
+      if (!ids) throw new Error('Can not create ContainerSelection: container "' + containerPath + '" does not exist.')
+      let start = _normalizeCoor(doc, { path: data.startPath, offset: data.startOffset, containerPath })
+      let end = _normalizeCoor(doc, { path: data.endPath, offset: data.endOffset, containerPath })
+      if (!_isDefined(data.reverse)) {
+        if (compareCoordinates(doc, containerPath, start, end) > 0) {
+          [start, end] = [end, start]
+          data.reverse = true
+        }
+      }
+      sel = new ContainerSelection(containerPath, start.path, start.offset, end.path, end.offset, data.reverse, data.surfaceId)
+      break
+    }
+    case 'node': {
+      sel = createNodeSelection({
+        doc,
+        nodeId: data.nodeId,
+        mode: data.mode,
+        containerPath: data.containerPath,
+        reverse: data.reverse,
+        surfaceId: data.surfaceId
+      })
+      break
+    }
+    case 'custom': {
+      sel = CustomSelection.fromJSON(data)
+      break
+    }
+    default:
+      throw new Error('Illegal selection type', data)
+  }
+  if (!sel.isNull()) {
+    sel.attach(doc)
+  }
+  return sel
+}
+
+function _normalizeCoor (doc, { path, offset, containerPath }) {
+  // NOTE: normalizing so that a node coordinate is used only for 'isolated nodes'
+  if (path.length === 1) {
+    // FIXME: originally getContainerRoot was called here
+    // however in this case
+    const node = getContainerRoot(doc, containerPath, path[0])
+    if (node.isText()) {
+      // console.warn("DEPRECATED: don't use node coordinates for TextNodes. Use selectionHelpers instead to set cursor at first or last position conveniently.")
+      return new Coordinate(node.getPath(), offset === 0 ? 0 : node.getLength())
+    } else if (node.isList()) {
+      // console.warn("DEPRECATED: don't use node coordinates for ListNodes. Use selectionHelpers instead to set cursor at first or last position conveniently.")
+      if (offset === 0) {
+        const item = node.getItemAt(0)
+        return new Coordinate(item.getPath(), 0)
+      } else {
+        const item = doc.get(last(node.items))
+        return new Coordinate(item.getPath(), item.getLength())
+      }
+    }
+  }
+  return new Coordinate(path, offset)
 }
 
 export function createNodeSelection ({ doc, nodeId, containerPath, mode, reverse, surfaceId }) {
